Stern Gerlach Diagram
A “triple” Stern-Gerlach experiment.  When a beam of electrons is passed through an inhomogeneous magnetic field, the electrons are separated according to the component of their spin angular momentum in the direction of the field.  Here, the first magnet measures the angular momentum in the z direction, and forces each atom to “declare” itself as one of the eigenstates of  Sz.  But the second magnetic field measures the orbital angular momentum in the y direction, forcing the atom to declare itself as an eigenstate of Sz.  This “scrambles” the angular momentum about the z-axis, and so when we pass the electron beam through another Stern-Gerlach magnet aligned with the z-axis, we observe both eigenstates of Sz.